requirements.txt
python-dotenv
python-kafka
s3fs